requirements.txt
pandas
numpy
yfinance
scikit-learn
requests
python-dotenv
beautifulsoup4
